// src/common/enums/TimeFormat.ts
export enum TIME_FORMAT {
    HHMMA = 'hh:mm A',
    HMMA = 'h:mm A',
    HH_MM_SS = 'hh:mm:ss',
    DDMMMY = 'DD MMM, YYYY',
    YMMDD = 'YYYYMMDD',
    Y_MM_DD = 'YYYY-MM-DD',
    MMDDY = 'MM/DD/YYYY ',
    YMDHMS = 'YYYY-MM-DD HH:mm:ss',
    YMDTHMS = 'YYYY-MM-DDTHH:mm:ss',
    YMDTHMSZZ = 'YYYY-MM-DDTHH:mm:ssZZ',
    MMMDDYHMMA = 'MMM DD, YYYY  H:mm A',
    DDDDMMMDD = 'dddd, MMM DD',
    DDDDMMMDO = 'dddd, MMM Do',
    DDDDMMMMDO = 'dddd, MMMM Do',
    DMMMYYYYHHMMSS = 'D-MMM-YYYY HH:mm:ss',
    HHMMSS = 'HH:mm:ss',
    MMDD = 'MM/DD',
    MDD = 'M/DD',
    MDYY = 'M/D/YY',
    MD = 'M/D',
    YYYY = 'YYYY/',
    Y = 'YYYY',
    MMMDDY = 'MMM DD , YYYY',
    MMM_DD_YYYY = 'MMM DD, YYYY',
    MMMDDYHHMMA = 'MMM DD, YYYY  h:mm A',
    MMM_D_YYYY_HH_MM_A = 'MMM D, YYYY  hh:mm A',
    MMMDDYYYY = 'MMMM DD, YYYY',
    MM_DD_YYYY = 'MM/DD/YYYY',
    DDD_MMM_DD_YYYY = 'ddd, MMM DD, YYYY',
    DDD_MMM_D_YYYY = 'ddd, MMM D, YYYY',
    D = 'd',
    DD = 'DD',
    DDD = 'ddd',
    DDDD = 'dddd',
    Z = 'Z',
    W = 'w',
    SCHEDULE_DATE_FORMAT = 'MMMM D, YYYY',
    MMM_DD = 'MMM DD',
    MMMM_D = 'MMMM D',
    YMD_HMS = 'HHMMSS',
    HH_MM = 'HH:mm',
    MMM_D_YYYY = 'MMM D, YYYY',
    YYMMDDTHHMMSS = 'YYYYMMDDTHHmmss',
    YMDTHMSSZZ = 'YYYY-MM-DDTHH:mm:ss.SSSZZ',
    MONTH = 'YYYY-MM'
}